requirements.txt
Django>=4.0
pandas
liac-arff
matplotlib
seaborn
gunicorn